+++ common/src/game/projectile.c
@@ -62,9 +62,9 @@ void projectileNewC(const character *c, being target, uint style){
 	const float mx =  1.f - c->aimFade;
 	const float mz = -1.f;
 	vec pos = c->pos;
-	pos.x += ((cos((c->rot.yaw+90.f)*PI/180) * cos(c->rot.pitch*PI/180))*mz) + cos((c->rot.yaw)*PI/180)*mx;
-	pos.y += (sin(c->rot.pitch*PI/180)*mz);
-	pos.z += ((sin((c->rot.yaw+90.f)*PI/180) * cos(c->rot.pitch*PI/180))*mz) + sin((c->rot.yaw)*PI/180)*mx;
+	pos.x += ((cosf((c->rot.yaw+90.f)*PI/180) * cosf(c->rot.pitch*PI/180))*mz) + cosf((c->rot.yaw)*PI/180)*mx;
+	pos.y += (sinf(c->rot.pitch*PI/180)*mz);
+	pos.z += ((sinf((c->rot.yaw+90.f)*PI/180) * cosf(c->rot.pitch*PI/180))*mz) + sinf((c->rot.yaw)*PI/180)*mx;
 
 	const float inacc = MIN(96.f,(c->inaccuracy*0.2f)) / (1.f + (c->aimFade * c->zoomFactor));
 	const float yaw   = c->rot.yaw   + (rngValf()-0.5f)*inacc;
